More work on TripPurpose-base main script to do output
* added write method that doesn't need ODMatrix to have predefined VOTs
and modes (this will need more work long term as we transition to
built-in STA)

* VOTs are now Floats instead of doubles

+++ src/edu/utexas/wrap/ModelInput.java
@@ -201,5 +201,5 @@ public interface ModelInput {
     
     Collection<TimePeriod> getUsedTimePeriods();
     
-    double getVOT(TripPurpose purpose, MarketSegment segment, Mode mode);
+    Float getVOT(TripPurpose purpose, MarketSegment segment, Mode mode);
 }
